release v2.0
Таймер зворотнього відліку;
Автоматична відправка результатів торгів і чату;
Кожні X сек. підвищення на один крок;
Можливість ліцетатору писати в чат;
Відредаговані дії користувачів - Кнопки онлайна по різні сторони.
Фон кнопки офлайна градієнтом $offline;
Виділені №, остаточні ціну/вартість, крок;
Повідомлення адміна в чаті виділяються $bluegrey;
Великі повідомлення в чаті відображаються правильно;
Адмін: встановлення наступної сесії;
Адмін: зміна паролю;
Адмін: підвантаження шаблону заяявки;
Адмін: встановлення часу, відведеного на крок;
Адмін: встановлення пошт на які надсилаються системні повідомлення;
Окрім Верифікації, метод "допуску до торгів";
Валідація реєстраційних даних;
Інформативніше наповнення вкладки доступу до торгів;
Розділення сесій адміністратора і користувача;
Оновлення інформації про стан верифікації/допуску до торгів без
необхідності перезаходити до профілю;
Архів заявок з можливістю їх завантаження;

+++ src/admin/scripts/php/admin-delete-lot.php
@@ -9,23 +9,25 @@ session_start();
 if ($_SESSION["id"] !== 'ADMIN') {
     session_unset();
     session_destroy();
-    header('Location: index.html');
+    header('Location: /index.php');
     die();
 }
 mb_internal_encoding("UTF-8");
 
-$host = 'joncolab.mysql.ukraine.com.ua';
-$username = 'joncolab_saladin';
-$password = '2014';
-$db = 'joncolab_trade';
+if (isset($_POST["lotId"])) {
+    $host = 'joncolab.mysql.ukraine.com.ua';
+    $username = 'joncolab_saladin';
+    $password = '2014';
+    $db = 'joncolab_trade';
 
-$connection = new mysqli($host, $username, $password, $db);
-if ($connection->connect_error) {
-    die('База даних не може опрацювати запит зараз, спробуйте за кілька хвилин');
+    $connection = new mysqli($host, $username, $password, $db);
+    if ($connection->connect_error) {
+        die('База даних не може опрацювати запит зараз, спробуйте за кілька хвилин');
+    }
+    $id = $_POST["lotId"];
+    $connection->set_charset('utf8');
+    $sql = "DELETE FROM lots WHERE id='" . $id . "'";
+    $connection->query($sql);
+    $connection->close();
 }
-$id = $_POST["lotId"];
-$connection->set_charset('utf8');
-$sql = "DELETE FROM lots WHERE id='" . $id . "'";
-$connection->query($sql);
-$connection->close();
 exit();
